convert DatePicker to ts
