--- src/DatePicker/index.d.ts
@@ -1,2 +1,0 @@
-import DatePicker from './DatePicker';
-export default DatePicker;
