--- src/DatePicker/index.js
@@ -1,2 +1,0 @@
-import DatePicker from './DatePicker';
-export default DatePicker;
