+++ src/DatePicker/index.ts
@@ -1,0 +1,3 @@
+import DatePicker, { DatePickerProps } from './DatePicker';
+export default DatePicker;
+export { DatePickerProps };
