--- typings/moment.d.ts
@@ -1,7 +1,0 @@
-declare module 'moment' {
-  namespace moment {
-    type Moment = any
-  }
-
-  export default moment;
-}
